requirements.txt
diffusers==0.30.3
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
Pillow==11.0.0
pytorch_fid==0.3.0
scipy==1.14.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
